code/retentionboards_retentioneering/app/requirements.txt
django
gunicorn
psycopg2-binary
redis
celery
kombu
retentioneering
django-redis
umap-learn
Pillow
